# server/swagger/cada.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "CHoRUS Equitable AI CADA API",
      "version": "1.0.0"
    },
...
    "paths": {
...
      "/cada/event/annotationValue": {
        "post": {
          "tags": [
            "cadaEvent"
          ],
          "summary": "Create annotation value",
          "operationId": "createAnnotationValue",
          "parameters": [],
          "requestBody": {
            "$ref": "#/components/requestBodies/cadaAnnotationValue"
          },
          "responses": {
            "200": {
              "description": "Success"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/cada/event/adjudicationValue": {
        "post": {
          "tags": [
            "cadaEvent"
          ],
          "summary": "Create adjudication value",
          "operationId": "createAdjudicationValue",
          "parameters": [],
          "requestBody": {
            "$ref": "#/components/requestBodies/cadaAdjudicationValue"
          },
          "responses": {
            "200": {
              "description": "Success"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
    },
    "components": {
...
      "requestBodies": {
...
        "cadaAnnotationValue": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "field": {
                    "type": "string",
                    "default": "Afib"
                  },
                  "value": {
                    "type": "string",
                    "default": "not sure"
                  },
                  "cadaAnnotationId": {
                    "type": "integer",
                    "default": 0
                  }
                }
              }
            }
          }
        },
        "cadaAdjudicationValue": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "field": {
                    "type": "string",
                    "default": "Afib"
                  },
                  "value": {
                    "type": "string",
                    "default": "not sure"
                  },
                  "userId": {
                    "type": "integer",
                    "default": 0
                  },
                  "cadaEventId": {
                    "type": "integer",
                    "default": 0
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }